Table_extraction/requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
keyboard
setuptools
#glob
#warnings
matplotlib>=3.2.2
pandas
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
#torch>=1.7.0
#torchvision>=0.8.1
#tqdm>=4.41.0
seaborn>=0.11.0

# pygaze --------------------------------------
#pylink>=0.3.2
#PyLink
python-pygaze

# ocr------------------------------------------
pybind11
#fastwer
pytesseract

# logging -------------------------------------
#tensorboard>=2.4.1
# wandb

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization
# tensorflow==2.4.1  # for TFLite export

# extras --------------------------------------
#Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# albumentations>=1.0.3
#thop  # FLOPs computation

# ocr development ---------------------------
#wheel
#pydot
#graphviz
